Bitwise access using Literals (#275)
* Bitaccess lexing

* Parse bitaccess

* Add error reporting tests

* Add validation support for bitwise operations

* Codegen for direct access

* Add correctness test and docu

Changed literal int size to i128 to fit the upper range of u64

+++ src/index/visitor.rs
@@ -255,7 +255,7 @@ fn visit_data_type(index: &mut Index, type_declatation: &UserTypeDeclaration) {
                     v,
                     enum_name.as_str(),
                     Some(ast::AstStatement::LiteralInteger {
-                        value: i as i64,
+                        value: i as i128,
                         location: SourceRange::undefined(),
                         id: 0,
                     }),
